Moved rooms & voting into a route file.
